introduce ClassSwappingRule
